Derive the package prefix from MINGW_PACKAGE_PREFIX everywhere
The "!= MINGW*" was likely there to cover the MSYS env, back when there
only was MINGW32/64. So that is equal to an empty prefix.

For UCRT/CLANG this means it will expand names to mingw packages
names like in MINGW32/64, but hardcoded to MINGW32/64, which is
probably not what we want, but will be fixed in the next commit.

This makes pacboy behave the same under all environments again.

See #4