Default to the environment package name instead of mingw32/64
"pacboy -S ffmpeg" would by default install ffmpeg for mingw32 and mingw64.
Due to other logic bugs fixed in the last commit this did not apply
to other environments, where it would fail because no prefix would be added.

Ideally the command would work the same in all environments and would only
install for the active environment unless specifically asked for.

This changes things so that "pacboy -S ffmpeg" works the same as
"pacboy -S ffmpeg:p" did before in all envs.

This is a slight breaking change, though I'd argue the current behaviour was
unexpected and unlikely to be wanted by users.

Fixes #4